tests/wpt/tests/css/css-typed-om/the-stylepropertymap/properties/animation-duration.html

<meta charset="utf-8">
<title>'animation-duration' property</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

function assert_is_equal_with_clamping_time(input, result) {
  const time = input.to('s');

  if (time.value < 0)
    assert_style_value_equals(result, new CSSUnitValue(0, 's'));
  else
    assert_style_value_equals(result, time);
}

runListValuedPropertyTests('animation-duration', [
  {
    syntax: '<time>',
    specified: assert_is_equal_with_range_handling,
    computed: assert_is_equal_with_clamping_time
  },
]);

test((t) => {
  let div = document.createElement('div');
  t.add_cleanup(() => {
    div.remove();
  });
  document.body.append(div);
  let actual = div.computedStyleMap().get('animation-duration').toString();
  assert_equals(actual, 'auto');
}, 'Computed value of animation-duration is auto');

</script>
